I now added the modal displaying all of the available information
We open the modal when we click on the image. It displays the information for the item and we can also now hook it up to a database. So we can favorite the skins that we want to save and possibly email the person when they show up in the shop.