I made the ability to sign out and email if a skin is in the shop
We can now press a button that emails the current user that a skin they have favorited is currently in the shop. I also fixed the modal popping up when the page starts. I also made it so the button in the top right of the account page signs you out.